+++ src/components/ConfirmationCodeInput/ConfirmationCodeInput.js
@@ -1,18 +1,17 @@
 // @flow
+import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
 import { View, TextInput as TextInputNative, StyleSheet } from 'react-native';
-import PropTypes from 'prop-types';
 
-import { getClassStyle, getInputSpaceStyle, isMatchingCode } from '../../utils';
 import TextInput from '../TextInput';
+import { getClassStyle, getInputSpaceStyle } from '../../utils';
 import { getContainerStyle, styles } from './styles';
 import { validateCompareCode, validateInputProps } from './validation';
 
-import type { SyntheticEvent } from 'react-native/Libraries/Types/CoreEventTypes';
-import type { VariantNames, INDEX } from '../../types';
-import type { Props, State } from './types';
+import type { Props, State, KeyPressEvent } from './types';
+import type { VariantNames } from '../../types';
 
-// eslint-disable-next-line
+// eslint-disable-next-line no-use-before-define
 type DP = typeof ConfirmationCodeInput.defaultProps;
 
 const getDefaultCodeSymbols: number => Array<string> = codeLength =>
@@ -22,7 +21,55 @@ export default class ConfirmationCodeInput extends PureComponent<
   $Diff<Props, DP> & $Shape<DP>,
   State,
 > {
+  static propTypes = {
+    activeColor: PropTypes.string,
+    autoFocus: PropTypes.bool,
+    cellBorderWidth: PropTypes.number,
+    codeLength: PropTypes.number,
+    containerProps: PropTypes.object,
+    defaultCode: validateCompareCode,
+    inputProps: PropTypes.func,
+    inputStyle: PropTypes.func,
+    inactiveColor: PropTypes.string,
+    inputPosition: PropTypes.oneOf(['center', 'left', 'right', 'full-width']),
+    onChangeCode: PropTypes.func,
+    onFulfill: PropTypes.func.isRequired,
+    size: PropTypes.number,
+    space: PropTypes.number,
+    variant: PropTypes.oneOf([
+      'border-box',
+      'border-circle',
+      'border-b',
+      'border-b-t',
+      'border-l-r',
+      'clear',
+    ]),
+    keyboardType: TextInputNative.propTypes.keyboardType,
+    maskSymbol: PropTypes.string,
+  };
+
+  static defaultProps = {
+    activeColor: '#fff',
+    autoFocus: false,
+    cellBorderWidth: 1,
+    codeLength: 5,
+    containerProps: {},
+    defaultCode: null,
+    inputProps: null,
+    inputStyle: null,
+    inactiveColor: '#ffffff40',
+    inputPosition: 'center',
+    onChangeCode: null,
+    size: 40,
+    space: 8,
+    variant: 'border-box',
+    keyboardType: 'default',
+    maskSymbol: null,
+    canPasteCode: false,
+  };
+
   styles: Object;
+
   ignoreOnFocusHandler: boolean = false;
 
   constructor(...args: any) {
@@ -69,7 +116,7 @@ export default class ConfirmationCodeInput extends PureComponent<
 
   detectFirstFocus: boolean = false;
 
-  handlerOnFocus = (index: INDEX) => {
+  handlerOnFocus = (index: number) => {
     if (this.ignoreOnFocusHandler) {
       return;
     }
@@ -119,109 +166,94 @@ export default class ConfirmationCodeInput extends PureComponent<
     });
   }
 
-  static IOS_MINIMAL_DELAY = 20;
-
-  lastKeyEventTimestamp: number = 0;
-
-  onKeyPress = (e: SyntheticEvent<*>) => {
-    if (e.nativeEvent.key === 'Backspace') {
-      /**
-       * Due to a bug in RN iOS TextInput implementation, an unwanted backspace
-       * event is fired on key press after clearing out the text input.
-       * Typically this backspace event follows the actual event
-       * and the time duration is under 10ms.
-       * Added a check to see if we receive a backspace event under ~20ms
-       * after the last key press event. If found, do nothing.
-       *
-       * Bug link: https://github.com/facebook/react-native/issues/18374
-       */
-      if (
-        Math.abs(this.lastKeyEventTimestamp - e.timeStamp) <
-        ConfirmationCodeInput.IOS_MINIMAL_DELAY
-      ) {
-        return;
-      }
-
-      const { currentIndex } = this.state;
-      const nextIndex = currentIndex > 0 ? currentIndex - 1 : 0;
-
-      this.setFocus(nextIndex);
-    } else {
-      // Record non-backspace key event time stamp
-      this.lastKeyEventTimestamp = e.timeStamp;
+  onKeyPress = (e: KeyPressEvent) => {
+    if (e.nativeEvent.key !== 'Backspace') {
+      return;
     }
+    const { currentIndex } = this.state;
+    const nextIndex = currentIndex > 0 ? currentIndex - 1 : 0;
+
+    this.setFocus(nextIndex);
   };
 
-  isLastIndex(index: INDEX): boolean {
+  isLastIndex(index: number): boolean {
     return index === this.props.codeLength - 1;
   }
 
-  onInputCode = (character: string, index: INDEX) => {
+  normalizeNewCode(array: Array<string>): Array<string> {
+    const { codeLength } = this.props;
+
+    const clearArray = array.filter(Boolean);
+
+    // Slice code when user paste long text
+    return new Array(codeLength).fill('').map((e, i) => {
+      if (clearArray[i]) {
+        return clearArray[i];
+      }
+
+      return e;
+    });
+  }
+
+  getCurrentIndex(symbols: Array<string>): number {
+    // Try holes in the array [1,2,3,'']
+    const index = symbols.findIndex(symbol => !symbol);
+
+    if (index === -1) {
+      return this.props.codeLength - 1;
+    }
+
+    return index - 1;
+  }
+
+  handlerOnChangeText = (text: string, index: number) => {
     // Fix for react-native-web
     // Skip onChange when text deleted, onKeyPress must handled the behavior
-
-    if (!character) {
+    if (!text) {
       return;
     }
 
-    // On Android: Text code is filled very slowly https://github.com/facebook/react-native/commit/5017b86
-    if (!this.isLastIndex(index)) {
+    const { codeSymbols } = this.state;
+
+    if (!this.props.canPasteCode) {
+      text = text[0];
+    }
+
+    const newCodeSymbols = this.normalizeNewCode([...codeSymbols, ...text]);
+    const currentIndex = this.getCurrentIndex(newCodeSymbols);
+
+    this.setState({
+      codeSymbols: newCodeSymbols,
+      currentIndex: currentIndex + 1,
+    });
+
+    if (this.isLastIndex(currentIndex)) {
+      const { onFulfill } = this.props;
+      const code = newCodeSymbols.join('');
+
+      this.blur(index);
+
+      onFulfill(code);
+    } else {
       // Skip processing onFocus that changes state
       this.ignoreOnFocusHandler = true;
 
-      this.setFocus(this.state.currentIndex + 1);
+      this.setFocus(currentIndex + 1);
     }
-
-    const {
-      onFulfill,
-      compareWithCode,
-      ignoreCaseWhenCompareCode,
-    } = this.props;
-    const { currentIndex, codeSymbols } = this.state;
-    const newCodeSymbols = [...codeSymbols];
-
-    newCodeSymbols[index] = character;
-
-    if (this.isLastIndex(index)) {
-      const code = newCodeSymbols.join('');
-
-      if (compareWithCode) {
-        const isMatching = isMatchingCode(
-          code,
-          compareWithCode,
-          ignoreCaseWhenCompareCode,
-        );
-
-        onFulfill(code, isMatching);
-
-        if (isMatching) {
-          this.clear();
-        }
-      } else {
-        onFulfill(code);
-      }
-
-      this.blur(currentIndex);
-    }
-
-    this.setState(prevState => ({
-      codeSymbols: newCodeSymbols,
-      currentIndex: prevState.currentIndex + 1,
-    }));
   };
 
   codeInputRefs: Array<{ blur: () => void, focus: () => void }> = [];
 
-  setFocus(index: INDEX) {
+  setFocus(index: number) {
     this.codeInputRefs[index].focus();
     this.ignoreOnFocusHandler = false;
   }
 
-  blur(index: INDEX) {
+  blur(index: number) {
     this.codeInputRefs[index].blur();
   }
 
-  setInputRef = (ref: any, idx: INDEX) => {
+  setInputRef = (ref: any, idx: number) => {
     this.codeInputRefs[idx] = ref;
   };
 
@@ -231,18 +263,18 @@ export default class ConfirmationCodeInput extends PureComponent<
     return value ? maskSymbol || value.toString() : '';
   }
 
-  renderInput(value: string, index: INDEX) {
+  renderInput(value: string, index: number) {
     const {
-      getInputStyle,
+      inputStyle,
       autoFocus,
       variant,
       activeColor,
-      getInputProps,
+      inputProps,
       keyboardType,
     } = this.props;
     const { currentIndex } = this.state;
 
-    const customInputProps = getInputProps ? getInputProps(index) : null;
+    const customInputProps = inputProps ? inputProps(index) : null;
 
     if (process.env.NODE_ENV !== 'production') {
       validateInputProps(customInputProps);
@@ -267,17 +299,25 @@ export default class ConfirmationCodeInput extends PureComponent<
           this.styles.input,
           this.getClassStyle(variant, currentIndex === index),
           customInputProps && customInputProps.style,
-          getInputStyle
-            ? getInputStyle(index, currentIndex === index, Boolean(value))
+          typeof inputStyle === 'function'
+            ? inputStyle(index, currentIndex === index, Boolean(value))
             : null,
         ]}
-        // fix for emoji  '😂'.length // 2
-        maxLength={finalValue ? finalValue.length : 1}
-        onChangeText={this.onInputCode}
+        maxLength={this.calculateMaxLength(finalValue)}
+        onChangeText={this.handlerOnChangeText}
         onFocus={this.handlerOnFocus}
         onKeyPress={this.onKeyPress}
       />
     );
+  }
+
+  calculateMaxLength(value: string): number {
+    if (this.props.canPasteCode) {
+      return this.props.codeLength;
+    }
+
+    // Fix for emoji '👲'.length === 2
+    return value ? value.length : 1;
   }
 
   render() {
@@ -298,52 +338,4 @@ export default class ConfirmationCodeInput extends PureComponent<
       </View>
     );
   }
-
-  static propTypes = {
-    activeColor: PropTypes.string,
-    autoFocus: PropTypes.bool,
-    cellBorderWidth: PropTypes.number,
-    codeLength: PropTypes.number,
-    compareWithCode: validateCompareCode,
-    containerProps: PropTypes.object,
-    defaultCode: validateCompareCode,
-    getInputProps: PropTypes.func,
-    ignoreCaseWhenCompareCode: PropTypes.bool,
-    inactiveColor: PropTypes.string,
-    inputPosition: PropTypes.oneOf(['center', 'left', 'right', 'full-width']),
-    onChangeCode: PropTypes.func,
-    onFulfill: PropTypes.func.isRequired,
-    size: PropTypes.number,
-    space: PropTypes.number,
-    variant: PropTypes.oneOf([
-      'border-box',
-      'border-circle',
-      'border-b',
-      'border-b-t',
-      'border-l-r',
-      'clear',
-    ]),
-    keyboardType: TextInputNative.propTypes.keyboardType,
-    maskSymbol: PropTypes.string,
-  };
-
-  static defaultProps = {
-    activeColor: '#fff',
-    autoFocus: false,
-    cellBorderWidth: 1,
-    codeLength: 5,
-    compareWithCode: null,
-    containerProps: {},
-    defaultCode: null,
-    getInputProps: null,
-    ignoreCaseWhenCompareCode: false,
-    inactiveColor: '#ffffff40',
-    inputPosition: 'center',
-    onChangeCode: null,
-    size: 40,
-    space: 8,
-    variant: 'border-box',
-    keyboardType: 'default',
-    maskSymbol: null,
-  };
 }