+++ src/components/ConfirmationCodeInput/types.js
@@ -1,13 +1,17 @@
 // @flow
+import type { SyntheticEvent } from 'react-native/Libraries/Types/CoreEventTypes';
 import type { KeyboardType } from 'react-native/Libraries/Components/TextInput/TextInput';
 import type { ViewProps } from 'react-native/Libraries/Components/View/ViewPropTypes';
-import type { VariantNames, INDEX, InputPositions } from '../../types';
+
+import type { VariantNames, InputPositions } from '../../types';
+
+export type KeyPressEvent = SyntheticEvent<{|
+  +key: string,
+|}>;
 
 export type Props = $ReadOnly<{|
   codeLength?: number,
   defaultCode?: string,
-  compareWithCode?: string,
-  ignoreCaseWhenCompareCode?: boolean,
 
   activeColor?: string,
   cellBorderWidth?: number,
@@ -20,17 +24,14 @@ export type Props = $ReadOnly<{|
   maskSymbol?: string,
 
   autoFocus?: boolean,
-  // isMatching will be true|false when "compareWithCode" prop isn't empty string
-  onFulfill: (code: string, isMatching: ?boolean) => void,
+  canPasteCode?: boolean,
+
+  onFulfill: (code: string) => void,
   onChangeCode?: (code: string) => void,
 
-  getInputProps?: (index: INDEX) => Object,
-  // help set custom style to any inputs
-  getInputStyle?: (
-    index: INDEX,
-    isFocused: boolean,
-    hasValue: boolean,
-  ) => Object,
+  inputProps?: (index: number) => Object,
+  // Help set a custom style to any inputs
+  inputStyle?: (index: number, isFocused: boolean, hasValue: boolean) => Object,
   containerProps?: ViewProps,
 
   testID?: any,